chore: simplified code by removing unnecessary component
